+++ cmd/heimdalld/service/init.go
@@ -1,4 +1,4 @@
-package main
+package service
 
 import (
 	"errors"
