+++ cmd/heimdalld/service/rollback.go
@@ -1,4 +1,4 @@
-package main
+package service
 
 import (
 	"fmt"
